Remove duplication by moving multi-select behaviour to the ListView component, ensure it also works in the widget editor
