+++ packages/edit-site/src/components/secondary-sidebar/list-view-sidebar.js
@@ -1,10 +1,7 @@
 /**
  * WordPress dependencies
  */
-import {
-	__experimentalListView as ListView,
-	store as blockEditorStore,
-} from '@wordpress/block-editor';
+import { __experimentalListView as ListView } from '@wordpress/block-editor';
 import { Button } from '@wordpress/components';
 import {
 	useFocusOnMount,
@@ -12,7 +9,7 @@ import {
 	useInstanceId,
 	useMergeRefs,
 } from '@wordpress/compose';
-import { useDispatch, useSelect } from '@wordpress/data';
+import { useDispatch } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import { closeSmall } from '@wordpress/icons';
 import { ESCAPE } from '@wordpress/keycodes';
@@ -24,58 +21,6 @@ import { store as editSiteStore } from '../../store';
 
 export default function ListViewSidebar() {
 	const { setIsListViewOpened } = useDispatch( editSiteStore );
-
-	const {
-		getBlockParents,
-		getBlockSelectionStart,
-		hasMultiSelection,
-		hasSelectedBlock,
-	} = useSelect( blockEditorStore );
-
-	const { clearSelectedBlock, multiSelect, selectBlock } = useDispatch(
-		blockEditorStore
-	);
-
-	async function selectEditorBlock( clientId, event ) {
-		if ( ! event.shiftKey ) {
-			await clearSelectedBlock();
-			selectBlock( clientId, -1 );
-		} else if ( event.shiftKey ) {
-			event.preventDefault();
-
-			if ( ! hasSelectedBlock() && ! hasMultiSelection() ) {
-				selectBlock( clientId, -1 );
-				return;
-			}
-
-			const blockSelectionStart = getBlockSelectionStart();
-
-			// By checking `blockSelectionStart` to be set, we handle the
-			// case where we select a single block. We also have to check
-			// the selectionEnd (clientId) not to be included in the
-			// `blockSelectionStart`'s parents because the click event is
-			// propagated.
-			const startParents = getBlockParents( blockSelectionStart );
-
-			if (
-				blockSelectionStart &&
-				blockSelectionStart !== clientId &&
-				! startParents?.includes( clientId )
-			) {
-				const startPath = [ ...startParents, blockSelectionStart ];
-				const endPath = [ ...getBlockParents( clientId ), clientId ];
-				const depth = Math.min( startPath.length, endPath.length ) - 1;
-				const start = startPath[ depth ];
-				const end = endPath[ depth ];
-
-				// Handle the case of having selected a parent block and
-				// then shift+click on a child.
-				if ( start !== end ) {
-					multiSelect( start, end );
-				}
-			}
-		}
-	}
 
 	const focusOnMountRef = useFocusOnMount( 'firstElement' );
 	const headerFocusReturnRef = useFocusReturn();
@@ -115,7 +60,6 @@ export default function ListViewSidebar() {
 				] ) }
 			>
 				<ListView
-					onSelect={ selectEditorBlock }
 					showNestedBlocks
 					__experimentalFeatures
 					__experimentalPersistentListViewFeatures
